requirements.txt
datasets==2.11.0
flet==0.6.2
qdrant_client==1.1.6
sentence_transformers
python-dotenv